+++ apps/web/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @stoker-platform/web-app
 
+## 0.5.212
+
+### Patch Changes
+
+- feat: improve Algolia search
+    - @stoker-platform/node-client@0.5.83
+    - @stoker-platform/utils@0.5.74
+    - @stoker-platform/web-client@0.5.87
+
 ## 0.5.211
 
 ### Patch Changes