+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @stoker-platform/cli
 
+## 0.5.130
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/types@0.5.62
+    - @stoker-platform/node-client@0.5.83
+    - @stoker-platform/utils@0.5.74
+
 ## 0.5.129
 
 ### Patch Changes