+++ packages/node-client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stoker-platform/node-client
 
+## 0.5.83
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/types@0.5.62
+    - @stoker-platform/utils@0.5.74
+
 ## 0.5.82
 
 ### Patch Changes